+++ packages/ckb-sdk-utils/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.19.0](https://github.com/nervosnetwork/ckb-sdk-js/compare/v0.18.0...v0.19.0) (2019-08-27)
+
+
+### Features
+
+* **type:** set the value of hash_type to "data" and "type" ([36a5512](https://github.com/nervosnetwork/ckb-sdk-js/commit/36a5512))
+* **utils:** rename the method of lockScriptToHash to scriptToHash ([40cdbaa](https://github.com/nervosnetwork/ckb-sdk-js/commit/40cdbaa))
+
+
+### BREAKING CHANGES
+
+* **type:** set the value of hash_type to "data" and "type"
+* **utils:** rename the method of lockScriptToHash to scriptToHash
+
+
+
+
+
 # [0.18.0](https://github.com/nervosnetwork/ckb-sdk-js/compare/v0.17.1...v0.18.0) (2019-08-10)
 
 **Note:** Version bump only for package @nervosnetwork/ckb-sdk-utils